backend/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic[email]==2.10.6
